Modify to MetaSchema
