+++ core/src/db/meta/backend/MetaSchema.cpp
@@ -9,4 +9,4 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "db/meta/backend/MetaField.h"
+#include "db/meta/backend/MetaSchema.h"
