+++ core/src/db/meta/backend/MetaSchema.h
@@ -61,5 +61,64 @@ class MetaField {
     std::string setting_;
 };
 
+using MetaFields = std::vector<MetaField>;
+
+class MetaSchema {
+ public:
+    MetaSchema(const std::string& name, const MetaFields& fields) : name_(name), fields_(fields) {
+    }
+
+    std::string
+    name() const {
+        return name_;
+    }
+
+    std::string
+    ToString() const {
+        std::string result;
+        for (auto& field : fields_) {
+            if (!result.empty()) {
+                result += ",";
+            }
+            result += field.ToString();
+        }
+
+//        std::string constraints;
+//        for (auto& constraint : constraint_fields_) {
+//            if (!constraints.empty()) {
+//                constraints += ",";
+//            }
+//            constraints += constraint.name();
+//        }
+//
+//        if (!constraints.empty()) {
+//            result += ",constraint uq unique(" + constraints + ")";
+//        }
+
+        return result;
+    }
+
+    // if the outer fields contains all this MetaSchema fields, return true
+    // otherwise return false
+    bool
+    IsEqual(const MetaFields& fields) const {
+        std::vector<std::string> found_field;
+        for (const auto& this_field : fields_) {
+            for (const auto& outer_field : fields) {
+                if (this_field.IsEqual(outer_field)) {
+                    found_field.push_back(this_field.name());
+                    break;
+                }
+            }
+        }
+
+        return found_field.size() == fields_.size();
+    }
+
+ private:
+    std::string name_;
+    MetaFields fields_;
+};
+
 }
